win64 fixes

+++ orkige_engine/engine_fastgui/FastGuiSlider.cpp
@@ -210,7 +210,7 @@ namespace Orkige
 				if (dist < minDist)
 				{
 					minDist = dist;
-					closestIndex = i;
+					closestIndex = (int)i;
 				}
 			}
 			oAssert(closestIndex != -1);
